Send all requests through a Session instance
Commit in preparation for a PR of pywebhdfs to use requests.Session. This commit
is the equivalent of
https://github.com/pywebhdfs/pywebhdfs/pull/40/commits/107c303c1342eb1df69dde845efe372fe4c023cb
whose original message was:

This allows the requests library to make use of KeepAlive HTTP
sessions instead of making a new socket for each operation. Increadibly
useful when making large numbers of requests to the same namenode, can
easily lead to ephem port exhaustion.

Signed-off-by: David Ojeda <david.ojeda@mensiatech.com>